feat(drivers): Implement IO Handling
Create a driver to handle initializing the mcu and any pheripherials (e.g watchdog, clocks, etc).
Implement a UI to abstract and simplify the TI's library that handles the hardware.
Map the enums to actual pins on the board, add logic to configure and read each pin.
Test the firmware using test functions to use each pin as an output and input.