feat(build): verify hardware
Verify hardware by configuring pin 3.4 with a pull up resistor and test if it is high.
While there is an internal port 3 on the launchpad it lacks physical connection. Take
advantage of this to differentiate between the test board (launchpad) and the PCB (jr).
Test if it works with blink led test function, and update make file to handle the changes.