feature(io): Enable Interrupts
Write the code enabling, configuring, triggering, and handling an interrupt input.
This will help the microcontroller react to events in a timely manner instead of waiting
or missing them completely, however increases the code complexity. Make sure to enable
interrupts globally. Create a simple test to test the functionality. Increase blink speed
in the assert handler.